Support multiple root cpp build configurations
Added support for multiple root cpp build configurations.
Instead of maintaining a single active configuration, a map is
used to maintain the relationship between the workspace root and its
given `CppBuildConfiguration`. This feature is an experimental
PR based on new developments in clangd to support multiple
compilation databases.

By default, when selecting multiple build configurations, the extension
will create a merged compilation database. Added a preference to support
an experimental clangd feature (that might either change or be removed)
which would not generate this aggregated database, but rather just
notify clangd of all the databases the user wants to use.

Until this is officially supported you shouldn't set this experimental
preference, unless for testing purposes.

- Update the `cpp` manager to handle the new data structure.
- CppBuildConfigurationManager creates a merged compilation database
  when multiple configs are used.
- Update the test cases (fix the task test case with incorrect imports
  and false positive results).

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>
Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>